TextDetection/requirements.txt
imutils==0.5.3
numpy==1.17.4
opencv-python==4.1.2.30
Pillow==6.2.1
